requirements.txt
streamlit==1.38.0
streamlit-option-menu==0.3.13
langchain==0.2.16 
langchain-core==0.2.38 
langchain-text-splitters==0.2.4 
langchain_ollama==0.1.3 
langsmith==0.1.110 
lxml==5.3.0 
multidict==6.0.5 
numpy==1.26.4 
ollama==0.3.2 
orjson==3.10.7 
webencodings==0.5.1 
yarl==1.9.7
langchain
langchain-community
pypdf
faiss-cpu
faiss-gpu
langchain-google-genai
python-dotenv
asgiref==3.8.1 
backoff==2.2.1 
bcrypt==4.2.0 
build==1.2.2 
chroma-hnswlib==0.7.6 
chromadb==0.5.5 
coloredlogs==15.0.1 
flatbuffers==24.3.25 
humanfriendly==10.0 
importlib-metadata==8.4.0 
kubernetes==30.1.0 
langchainhub==0.1.21 
mmh3==4.1.0 
monotonic==1.6 
mpmath==1.3.0 
onnxruntime==1.19.2 
opentelemetry-api==1.27.0 
opentelemetry-exporter-otlp-proto-common==1.27.0 
opentelemetry-exporter-otlp-proto-grpc==1.27.0 
opentelemetry-instrumentation==0.48b0 
opentelemetry-instrumentation-asgi==0.48b0 
opentelemetry-instrumentation-fastapi==0.48b0 
opentelemetry-proto==1.27.0 
opentelemetry-sdk==1.27.0 
opentelemetry-semantic-conventions==0.48b0 
opentelemetry-util-http==0.48b0 
overrides==7.7.0 
posthog==3.6.5 
pypika==0.48.9 
pyproject_hooks==1.1.0 
pyreadline3==3.4.1 
regex==2024.7.24 
setuptools==74.1.2 
sympy==1.13.2 
tiktoken==0.7.0 
tokenizers==0.20.0 
types-requests==2.32.0.20240907 
zipp==3.20.1
pytube-15.0.0 
youtube-transcript-api==0.6.2
langchain-chroma==0.1.4